requirements.txt
anyio==4.2.0
APScheduler==3.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
python-telegram-bot==13.3
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
